requirements.txt
# Core Dependencies
openai>=1.0.0
whisper>=1.1.10
pyaudio>=0.2.13
numpy>=1.24.0
scipy>=1.10.0

# FastAPI for API endpoints
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Audio Processing
sounddevice>=0.4.6
librosa>=0.10.0
pydub>=0.25.1

# Natural Language Processing
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.2

# Database and Storage
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Context and Memory
chromadb>=0.4.15
langchain>=0.1.0

# Monitoring and Logging
prometheus-client>=0.19.0
structlog>=23.2.0
python-json-logger>=2.0.7

# Configuration Management
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Code Quality
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.13.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0
bcrypt>=4.1.0

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0
aiohttp>=3.9.0
tenacity>=8.2.0

# Optional: GPU Support
# torch-cuda>=2.1.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118

# Optional: Sensor Integration
# RPi.GPIO>=0.7.1  # For Raspberry Pi GPIO pins
# adafruit-circuitpython-dht>=3.7.0  # DHT temperature sensors
# adafruit-circuitpython-bme280>=2.6.0  # BME280 sensors
